+++ CHANGELOG.md
@@ -16,6 +16,8 @@ tiling after we first remove that requirement. This is added to
 multiple different obs and one requirement list of size possibly greater than one.
 Previously it was only doing the case where a single ob's factor is implied by a
 requirement.
+- `PointJumpingFactory` which adds rules where requirements and assumptions can be 
+swapped around a fusable row or column.
 
 ### Fixed
 - `ForgetTrackedSearcher` was not retroactively applying strategies that had a `basis`.